LPS-170680 Extract BasicInfo card to its own component
