+++ modules/apps/notification/notification-web/src/main/resources/META-INF/resources/js/components/Attachments.tsx
@@ -25,7 +25,6 @@ import React, {useEffect, useMemo, useState} from 'react';
 import {defaultLanguageId} from '../util/constants';
 
 import './Attachments.scss';
-import {TNotificationTemplate} from './EditNotificationTemplate';
 
 export function Attachments({setValues, values}: IProps) {
 	const [objectDefinitions, setObjectDefinitions] = useState<
@@ -210,6 +209,6 @@ export function Attachments({setValues, values}: IProps) {
 }
 
 interface IProps {
-	setValues: (values: Partial<TNotificationTemplate>) => void;
-	values: Partial<TNotificationTemplate>;
+	setValues: (values: Partial<NotificationTemplate>) => void;
+	values: Partial<NotificationTemplate>;
 }
