+++ modules/apps/notification/notification-web/src/main/resources/META-INF/resources/js/components/FreeMarkerTemplateEditor.tsx
@@ -22,15 +22,14 @@ import {createResourceURL, fetch} from 'frontend-js-web';
 import React, {useMemo, useState} from 'react';
 
 import {defaultLanguageId} from '../util/constants';
-import {TNotificationTemplate} from './EditNotificationTemplate';
 
 interface FreeMarkerTemplateEditorProps {
 	baseResourceURL: string;
 	objectDefinitions: ObjectDefinition[];
 	selectedLocale: Liferay.Language.Locale;
 	setSelectedLocale: (value: Liferay.Language.Locale) => void;
-	setValues: (values: Partial<TNotificationTemplate>) => void;
-	values: TNotificationTemplate;
+	setValues: (values: Partial<NotificationTemplate>) => void;
+	values: NotificationTemplate;
 }
 
 export function FreeMarkerTemplateEditor({
