+++ modules/apps/notification/notification-web/types/src/main/resources/META-INF/resources/js/components/Attachments.d.ts
@@ -15,10 +15,9 @@
 /// <reference types="react" />
 
 import './Attachments.scss';
-import {TNotificationTemplate} from './EditNotificationTemplate';
 export declare function Attachments({setValues, values}: IProps): JSX.Element;
 interface IProps {
-	setValues: (values: Partial<TNotificationTemplate>) => void;
-	values: Partial<TNotificationTemplate>;
+	setValues: (values: Partial<NotificationTemplate>) => void;
+	values: Partial<NotificationTemplate>;
 }
 export {};
